:recycle: Refactor  코드 수정
